Small [tm] refactoring: Move parts of the server that are either generically cardgame or protocol related to cardgame and protocol modules
