+++ Cargo.toml
@@ -1,6 +1,41 @@
 [workspace]
+resolver = "2"
+members = ["server", "client", "protocol", "playing_cards"]
 
-members = [
-    "server",
-    "client",
-]
+[workspace.lints.rust]
+unsafe_code = "deny"
+keyword_idents = "warn"
+missing_copy_implementations = "warn"
+non_ascii_idents = "warn"
+noop_method_call = "warn"
+unused_crate_dependencies = "warn"
+unused_extern_crates = "warn"
+unused_import_braces = "warn"
+
+[workspace.lints.clippy]
+cognitive_complexity = "warn"
+dbg_macro = "warn"
+debug_assert_with_mut_call = "warn"
+doc_link_with_quotes = "warn"
+doc_markdown = "warn"
+empty_line_after_outer_attr = "warn"
+empty_structs_with_brackets = "warn"
+float_cmp = "warn"
+float_cmp_const = "warn"
+float_equality_without_abs = "warn"
+missing_const_for_fn = "warn"
+missing_errors_doc = "warn"
+missing_panics_doc = "warn"
+mod_module_files = "warn"
+option_if_let_else = "warn"
+print_stderr = "warn"
+semicolon_if_nothing_returned = "warn"
+unseparated_literal_suffix = "warn"
+similar_names = "warn"
+suspicious_operation_groupings = "warn"
+unused_self = "warn"
+use_debug = "warn"
+used_underscore_binding = "warn"
+useless_let_if_seq = "warn"
+wildcard_dependencies = "warn"
+wildcard_imports = "warn"
