+++ client/Cargo.toml
@@ -3,6 +3,8 @@ name = "client"
 version = "0.1.0"
 edition = "2021"
 
+[lints]
+workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
@@ -12,8 +14,8 @@ clap = { version = "4.3.10", features = [
 ], default-features = false }
 tokio = { version = "1.29.1", features = [
     "macros",
-    "io-util",
+    "io-std",
     "net",
     "rt-multi-thread",
-    "io-std",
+    "io-util",
 ], default-features = false }
