+++ playing_cards/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "playing_cards"
+version = "0.1.0"
+edition = "2021"
+
+[lints]
+workspace = true
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+itertools = "0.10.5"
+rand = { features = [
+    "std_rng",
+    "std",
+], default-features = false, version = "0.8.5" }
+serde = { version = "1.0.160", features = [
+    "serde_derive",
+], default-features = false }
+
+[dev-dependencies]
+proptest = "1.1.0"
+proptest-derive = "0.3.0"
