+++ protocol/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "protocol"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lints]
+workspace = true
+
+[dependencies]
+itertools = "0.10.5"
+serde = { version = "1.0.160", features = [
+    "serde_derive",
+], default-features = false }
+playing_cards = { path = "../playing_cards" }
